backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
passlib[bcrypt]
PyJWT